feat(selectable menu)!: replace memorization required with selectable…

+++ modules/handler.py
@@ -11,45 +11,42 @@ def handle_empty_name(name: str) -> None:
         sys.exit(1)
 
 
-def handle_answers() -> None:
-    choice: str = input("Choice: ")
-
-    match choice:
-        case "1a":
-            dir_name: str = input("Folder name(s) to add (separated by space): ")
-            create_dir(dir_name)
-        case "1b":
-            file_name: str = input("File name(s) to add (separated by space): ")
-            create_file(file_name)
-        case "2a":
-            dir_name: str = input("Folder name(s) to delete (separated by space): ")
-            remove_dir(dir_name)
-        case "2b":
-            file_name: str = input("File name(s) to delete (separated by space): ")
-            remove_file(file_name)
-        case "3":
-            dir_name: str = input("Empty folder name(s) to delete (separated by space): ")
-            remove_empty_dir(dir_name)
-        case "4":
-            existing_dir: str = input("Existing folder name: ")
-            change_cd(existing_dir)
-        case "5a":
-            source: str = input("Folder name to move: ")
-            target: str = input("Folder name to move to: ")
-            move_dir(source, target)
-        case "5b":
-            source: str = input("File name(s) to move (separated by space): ")
-            target: str = input("Folder name to move file to: ")
-            move_file(source, target)
-        case "6":
-            old_name: str = input("Current folder or file name: ")
-            new_name: str = input("New folder or file name: ")
-            rename(old_name, new_name)
-        case "7":
-            dir_name: str = input("Folder name to add and go to: ")
-            create_change_dir(dir_name)
-        case "99":
-            print("You quit program.")
-            sys.exit(1)
-        case _:
-            print("Invalid option! Please try again!\n")
+def handle_answers(options: list[str], choice) -> None:
+    if choice == options[0]:
+        dir_name: str = input("Folder name(s) to add (separated by space): ")
+        create_dir(dir_name)
+    if choice == options[1]:
+        file_name: str = input("File name(s) to add (separated by space): ")
+        create_file(file_name)
+    if choice == options[2]:
+        dir_name: str = input("Folder name(s) to delete (separated by space): ")
+        remove_dir(dir_name)
+    if choice == options[3]:
+        file_name: str = input("File name(s) to delete (separated by space): ")
+        remove_file(file_name)
+    if choice == options[4]:
+        dir_name: str = input("Empty folder name(s) to delete (separated by space): ")
+        remove_empty_dir(dir_name)
+    if choice == options[5]:
+        existing_dir: str = input("Existing folder name: ")
+        change_cd(existing_dir)
+    if choice == options[6]:
+        source: str = input("Folder name to move: ")
+        target: str = input("Folder name to move to: ")
+        move_dir(source, target)
+    if choice == options[7]:
+        source: str = input("File name(s) to move (separated by space): ")
+        target: str = input("Folder name to move file to: ")
+        move_file(source, target)
+    if choice == options[8]:
+        old_name: str = input("Current folder or file name: ")
+        new_name: str = input("New folder or file name: ")
+        rename(old_name, new_name)
+    if choice == options[9]:
+        dir_name: str = input("Folder name to add and go to: ")
+        create_change_dir(dir_name)
+    if choice == options[10]:
+        print("You quit program.")
+        sys.exit(1)
+    
+    print("Invalid option! Please try again!\n")
